requirements.txt
numpy
scipy
pandas
matplotlib
torch
torchvision
transformers
Pillow
fastapi
uvicorn
python-multipart
opencv-python
reportlab
jinja2
scikit-learn
pytest
pytest-cov
facenet-pytorch
timm
google-generativeai
python-dotenv
